+++ backend/src/test/java/com/ssafy/togeball/domain/league/entity/ClubStadiumTest.java
@@ -22,8 +22,8 @@ public class ClubStadiumTest {
 
         //테스트 데이터 생성
         List<Club> clubs = new ArrayList<>();
-        Club club1 = new Club("LG","트윈스","logo.png",(byte)1);
-        Club club2 = new Club("두산","베어스","logo.png",(byte)5);
+        Club club1 = new Club("LG","트윈스","logo.png", 1);
+        Club club2 = new Club("두산","베어스","logo.png", 5);
         clubs.add(club1);
         clubs.add(club2);
         Stadium stadium1 = new Stadium(clubs, "잠실");
